Modified problems 8,11,12 of Set 3

+++ src/Set_3_Descriptive/Problem_12_Set_3.java
@@ -3,7 +3,7 @@ package Set_3_Descriptive;
 import java.util.Scanner;
 
 public class Problem_12_Set_3 {
-    
+
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int N, X;
@@ -12,6 +12,6 @@ public class Problem_12_Set_3 {
         System.out.print("Already has number of candies, X: ");
         X = sc.nextInt();
         double candyPacketNo = (double) (N - X) / 4;
-        System.out.println("Minimum number of candy packets = " + Math.ceil(candyPacketNo));
+        System.out.println("Minimum number of candy packets = " + (int) Math.ceil(candyPacketNo));
     }
 }
